Minor: Move file to correct package
