+++ app/src/main/java/com/chess/clock/engine/TimeControlParser.java
@@ -7,6 +7,7 @@ import android.util.Log;
 
 import com.chess.clock.engine.stage.Stage;
 import com.chess.clock.engine.time.TimeControl;
+import com.chess.clock.engine.time.TimeControlWrapper;
 import com.chess.clock.engine.time.TimeIncrement;
 import com.chess.clock.engine.time.TimeIncrementDelay;
 import com.chess.clock.engine.time.TimeIncrementFischer;
